Swap project order and update styles on homepage
Swapped the order and content of 'OmniMaze Auto' and 'The PI Line' projects on the homepage, updated their descriptions, icons, and action buttons. Increased the .project-title font size in styles.css from 37px to 42px for improved visibility and clarity that those are in fact headers and not just paragraph text.